Attempt to fix sporadic missing method deactive_embargo! errors
Refactors this monkey patch into a concern.
